Add support for multiple PCM music frequencies

+++ FASTDOOM/ns_multi.c
@@ -984,7 +984,7 @@ void MV_StopPlayback(
 int MV_PlayRaw(
     unsigned char *ptr,
     unsigned long length,
-    unsigned rate,
+    unsigned long rate,
     int vol,
     int left,
     int right,
